requirements.txt
djangorestframework==3.15.2
djangorestframework_simplejwt==5.3.1
PyJWT